+++ src/inline.js
@@ -1,5 +1,17 @@
-import Form from './form'
+import React from 'react'
+import useTemplate from './use-template'
 
-const InlineForm = Form
+function InlineForm({
+  template,
+  ...props
+}) {
+  const { options, Template } = useTemplate(template)
+  return (
+    <>
+      <script src="https://f.convertkit.com/ckjs/ck.5.js"></script>
+      <Template {...props} options={options} />
+    </>
+  )
+}
 
 export default InlineForm