+++ src/templates/mills-inline.js
@@ -16,10 +16,10 @@ function MillsInline({
   stacked,
 }) {
   return (
-    <form action={action} className="seva-form formkit-form" method="post" target={newTab ? '_blank' : '_self'} data-sv-form={formId} data-uid={options.uid} data-format="inline" data-version={options.version} data-options={options} min-width="400 500 600 700 800" style={{ backgroundColor: 'rgb(249, 250, 251)', borderRadius: '4px' }}>
+    <form action={action} className="seva-form formkit-form" method="post" target={newTab ? '_blank' : '_self'} data-sv-form={formId} data-uid={options.uid} data-format="inline" data-version={options.version} data-options={JSON.stringify(options)} min-width="400 500 600 700 800" style={{ backgroundColor: 'rgb(249, 250, 251)', borderRadius: '4px' }}>
       <div className="formkit-background" style={{ opacity: 0.2 }}></div>
       <div data-style="minimal">
-        <div className="formkit-header" data-element="header" style={{ color: 'rgb(77, 77, 77)', fontSize: '27px', fontWeight: 700 }}>
+        <div className="formkit-header" data-element="header" style={{ color: 'rgb(77, 77, 77)', fontSize: '27px', fontWeight: 700, textAlign: 'center' }}>
           <h1>Join the Newsletter</h1>
         </div>
         <div className="formkit-subheader" data-element="subheader" style={{ color: 'rgb(104, 104, 104)', fontsize: '18px' }}>Subscribe to get our latest content by email.</div>