+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Convertkit React Changelog
 
+### [v1.3.0](https://github.com/ConvertKit/convertkit-react/releases/tag/v1.3.0)
+
+- Add Powell and Charlotte inline templates
+- Modify build to inject only a single style tag
+- Fix conflicting styles across templates
+
+
 ### [v1.2.1](https://github.com/ConvertKit/convertkit-react/releases/tag/v1.2.1)
 
 - Ensure formId is being passed to templates